Added a comment_status field with choices.
The Choices status field is deliberately configured to use 0 for
disabled comments and 1 or 2 for enabled and frozen comments.
This allows you to decide if comments should be display with a 
simple if story.comment_status. 

The comments_frozen property was added to allow a descriptive way to test if the comment entry form should be shown.